[extension/sumologic] move api and credentials packages to internal
